gui updates
Two files for the calendar gui and a new event dialog box. Still need to
clean up a lot of the code, and implement a lot of functionality.
Currently, you can create an event in the dialog and it will print the
new event info in the command line. No graphical representation as of
yet.